Update tests, pool and worker

+++ src/Worker.php
@@ -12,15 +12,34 @@ class Worker
     const TERMINATING = 3;
     const TERMINATED = 4;
 
+    // shared preferences
+
+    protected $useSignals = false;
+
     // parent properties
     public $state = self::IDLE;
     protected $pid;
     protected $toChild;
 
+    /**
+     * Time between checks for new payload from parent
+     */
+    public $checkMicroTime = 100000;
+
     // child properties
     protected $parentPid;
     protected $toParent;
     protected $termAccepted = false;
+
+    /**
+     * Configures worker
+     * @param null|WorkerPool If passed, worker will use signals to communicate with pool
+     */
+    public function __construct($workerPool = null)
+    {
+        if ($workerPool instanceof WorkerPool)
+            $this->useSignals = true;
+    }
 
     public function start()
     {
@@ -30,8 +49,15 @@ class Worker
 
         $this->pid = $this->fork([$this, 'onWorkerStart'], [$sockets[1], $pid]);
         $this->toChild = $sockets[0];
+        socket_set_nonblock($this->toChild);
     }
 
+    /**
+     * Stops a worker and ends the process
+     * @param boolean $wait If true, this method will hold execution till the process end
+     * @return null|boolean Null, if worker is not active. False, if signal can't be sent to worker.
+     * True, if signal sent OR signal sent and worker terminated (if $wait = true).
+     */
     public function stop($wait = false)
     {
         if (!is_dir('/proc/'.$this->pid))
@@ -39,7 +65,8 @@ class Worker
 
         $this->state = self::TERMINATING;
 
-        posix_kill($this->pid, SIGTERM);
+        if (!posix_kill($this->pid, SIGTERM))
+            return false;
 
         if ($wait) {
             pcntl_waitpid($this->pid, $status);
@@ -48,8 +75,15 @@ class Worker
             if (!is_dir('/proc/'.$this->pid))
                 $this->state = self::TERMINATED;
         }
+        return true;
     }
 
+    /**
+     * Stops a worker and kills the process
+     * @param boolean $wait If true, this method will hold execution till the process end
+     * @return null|boolean Null, if worker is not active. False, if signal can't be sent to worker.
+     * True, if signal sent OR signal sent and worker terminated (if $wait = true).
+     */
     public function kill($wait = false)
     {
         if (!is_dir('/proc/'.$this->pid))
@@ -57,7 +91,8 @@ class Worker
 
         $this->state = self::TERMINATING;
 
-        posix_kill($this->pid, SIGKILL);
+        if (!posix_kill($this->pid, SIGKILL))
+            return false;
 
         if ($wait) {
             pcntl_waitpid($this->pid, $status);
@@ -66,10 +101,13 @@ class Worker
 
             if (!is_dir('/proc/'.$this->pid))
                 $this->state = self::TERMINATED;
-
         }
+        return true;
     }
 
+    /**
+     * @return boolean Returns whether worker is running right now.
+     */
     public function isRunning()
     {
         return $this->state == self::RUNNING;
@@ -77,27 +115,40 @@ class Worker
 
     public function checkForFinish()
     {
-        if (strlen($msg_size_bytes = socket_read($this->toChild, 4)) > 0) {
+        if (strlen($msg_size_bytes = socket_read($this->toChild, 4)) === 4) {
             $data = unpack('N', $msg_size_bytes);
             $msg_size = $data[1];
+            // echo 'Msg size: '.$msg_size.PHP_EOL;
 
             $msg = socket_read($this->toChild, $msg_size);
             $data = unserialize($msg);
+            // echo 'Msg: '.print_r($msg, true).PHP_EOL;
 
             // mark as idle
             $this->state = self::IDLE;
+            return true;
         }
+        return null;
     }
 
+    /**
+     * @return null|boolean
+     */
     public function checkForTermination()
     {
         $pid = pcntl_waitpid($this->pid, $status, WNOHANG);
-        var_dump($pid);
         if ($pid == $this->pid) {
             $this->state = self::TERMINATED;
+            socket_close($this->toChild);
+            return true;
         }
+        return null;
     }
 
+    /**
+     * Sends payload to worker
+     * @param array $data
+     */
     public function sendPayload($data)
     {
         $this->state = self::RUNNING;
@@ -108,6 +159,9 @@ class Worker
         socket_write($this->toChild, $data);
     }
 
+    /**
+     * @return integer The process ID of the worker
+     */
     public function getPid()
     {
         return $this->pid;
@@ -117,15 +171,28 @@ class Worker
     /// Child methods below only
     ///
 
+    /**
+     * The first method that is being called when the worker starts.
+     * @param socket $socket Socket to communicate with master process and transport payload / result.
+     * @param integer $parentPid The process ID of the parent
+     */
     protected function onWorkerStart($socket, $parentPid)
     {
+        declare(ticks = 1);
         // set signal handler
         pcntl_signal(SIGTERM, [$this, 'onSigTerm']);
 
         $this->toParent = $socket;
         $this->parentPid = $parentPid;
-        socket_set_block($this->toParent);
-        while (!$this->termAccepted && strlen($msg_size_bytes = socket_read($this->toParent, 4)) > 0) {
+        socket_set_nonblock($this->toParent);
+        $i = 0;
+        while (!$this->termAccepted) {
+            // echo ($i++).' try'.PHP_EOL;
+            // sleep for sometime
+            if (strlen($msg_size_bytes = socket_read($this->toParent, 4)) != 4) {
+                usleep($this->checkMicroTime);
+                continue;
+            }
             $data = unpack('N', $msg_size_bytes);
             $msg_size = $data[1];
 
@@ -141,19 +208,28 @@ class Worker
             socket_write($this->toParent, $data);
 
             // send SIGUSR1 to parent to indicate that's we've finished
-            posix_kill($this->parentPid, SIGUSR1);
+            if ($this->useSignals)
+                posix_kill($this->parentPid, SIGUSR1);
         }
-
+        socket_close($this->toParent);
         // notify we ended
-        posix_kill($this->parentPid, SIGUSR2);
+        if ($this->useSignals)
+            posix_kill($this->parentPid, SIGUSR2);
     }
 
+    /**
+     * The main handler and work executor in worker. Accepts all payload and should do all the work to process it.
+     */
     public function onPayload($abc)
     {
         echo 'I\'m just a worker with pid '.getmypid().'. Got payload: '.$abc.PHP_EOL;
         return 123;
     }
 
+
+    /**
+     * SIGTERM Handler to gracefully stop worker without losing any data.
+     */
     public function onSigTerm()
     {
         $this->termAccepted = true;
